Refactor package to make room for other commands
