+++ packages/kbn-check-saved-objects-cli/README.md
@@ -1,7 +1,5 @@
 # @kbn/check-saved-objects-cli
 
-Saved Objects CLI tool that can be used to check whether a snapshot of current
-mappings (i.e., mappings on main) is compatible with mappings we can extract
-from the current code.
+Saved Objects CLI tool that can be used to check validity / consistency of changes performed in Saved Objects definitions and mappings.
 
 See `node scripts/check_mappings_update --help` for more info.
