+++ packages/kbn-check-saved-objects-cli/package.json
@@ -2,6 +2,5 @@
   "name": "@kbn/check-saved-objects-cli",
   "private": true,
   "version": "1.0.0",
-  "license": "Elastic License 2.0 OR AGPL-3.0-only OR SSPL-1.0",
-  "main": "./src/run_check_mappings_update_cli"
+  "license": "Elastic License 2.0 OR AGPL-3.0-only OR SSPL-1.0"
 }
